requirements.txt
dj-database-url==0.4.1
Django==1.10.1
django-admin-sortable==2.0.19
django-dynamic-preferences==0.8.2
django-picklefield==0.3.2
gunicorn==19.6.0
persisting-theory==0.2.1
psycopg2==2.6.2
pytz==2016.7
requests==2.11.1
six==1.10.0
whitenoise==3.2.2
